+++ python/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `Person.item_ids` will now update dynamically when a new Person is created and/or relevant attributes of a Person change (i.e. `names`, `disable_name_matching`). (related to #7879)
+- `Person.item_ids` will now update dynamically when a NameSpecification's `name` or `id` attributes are modified.
 - Bugfix: NameSpecifications now prevent having unverified IDs set on them (#7901)
 - `PersonIndex.create` now deduplicates list of names (#7858)